# perf: Add mergeHistogram fast path