--- a/aggregators/merger.go
+++ b/aggregators/merger.go
@@ -385,7 +385,7 @@ func mergeSpanMetrics(to, from *aggregationpb.SpanMetrics) {
 // mergeHistogram merges two proto representation of HDRHistogram. The
 // merge assumes both histograms are created with same arguments and
 // their representations are sorted by bucket.
-// Caution: this function may mutate from.Count.
+// Caution: this function may mutate from.Counts.
 func mergeHistogram(to, from *aggregationpb.HDRHistogram) {
 	if len(from.Buckets) == 0 {
 		return